test: mappings and helpers
